backend/requirements.base.blended.txt
# ML
numpy==1.23.5
pandas==2.2.3
scikit_learn==1.6.1
scipy==1.15.1
torch==2.5.1
transformers==4.48.1
matplotlib==3.10.0
python-dotenv==1.0.1
pyyaml==6.0.1
catboost==1.2.7

# Testing
pytest==8.0.0

# Utils
tqdm==4.66.4

# Web
beautifulsoup4==4.12.3
requests==2.32.3

# Storage & Files
s3fs==2023.6.0
fsspec==2023.6.0
adlfs>=2023.6.0
gcsfs>=2023.6.0
boto3==1.28.17
mlflow==2.8.1
psycopg2-binary==2.9.9
